requirements.txt
black==23.3.0
certifi==2023.11.17
click==8.1.3
colorama==0.4.6
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==21.2.0
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pathspec==0.12.1
platformdirs==3.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.5
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.9.0 
wheel==0.42.0
whitenoise==5.3.0

python-dotenv==1.0.0
django==5.0.1 
asgiref==3.7.2
